*: add versioning to sdk and version constraint for generated project
